refactor(password-reset): decouple password reset flow from auth context
Move password reset actions, exceptions, requests, controllers,
and tests out of the auth context into a dedicated password-reset domain.

- Move password reset actions to Actions/PasswordReset
- Move password reset exceptions to dedicated namespaces
- Move password reset requests and controllers
- Rename ForgotPasswordAction to StartPasswordResetAction
- Reorganize feature and unit tests
- Update routes, validators, and translations
- Remove obsolete auth-coupled structure
