--- app/Http/Requests/Auth/ForgotPasswordRequest.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Auth;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class ForgotPasswordRequest extends FormRequest
-{
-    public function email(): string
-    {
-        return $this->input('email');
-    }
-
-    public function rules(): array
-    {
-        return [
-            'email' => ['required', 'string', 'email'],
-        ];
-    }
-}
