+++ app/Http/Requests/PasswordReset/StartPasswordResetRequest.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Http\Requests\PasswordReset;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class StartPasswordResetRequest extends FormRequest
+{
+    public function email(): string
+    {
+        return $this->input('email');
+    }
+
+    public function rules(): array
+    {
+        return [
+            'email' => ['required', 'string', 'email'],
+        ];
+    }
+}
