+++ lang/pt_BR/validators.php
@@ -4,10 +4,12 @@ return [
     'auth' => [
         'invalid_credentials' => 'Credenciais inválidas.',
         'email_not_exists' => 'O e-mail informado não existe.',
-        'password_reset_cooldown' => 'Usuário já solicitou um código para redefinição de senha.',
-        'invalid_password_reset_code' => 'Código de redefinição de senha é inválido ou não existe.',
-        'password_reset_attempts_exceeded' => 'Você excedeu o limite de tentativas de redefinição de senha. Tente novamente em alguns minutos',
-        'password_reset_blocked' => 'A redefinição de senha está temporariamente bloqueada para este usuário. Tente novamente em alguns minutos',
+    ],
+    'password_reset' => [
+        'cooldown' => 'Você já solicitou um código para redefinição de senha. Tente novamente em alguns minutos.',
+        'blocked' => 'A redefinição de senha está temporariamente bloqueada. Tente novamente em alguns minutos',
+        'invalid_code' => 'Código de redefinição de senha é inválido ou não existe.',
+        'attempts_exceeded' => 'Você excedeu o limite de tentativas de redefinição de senha. Tente novamente em alguns minutos',
     ],
     'device' => [
         'user_not_owner' => 'Somente o proprietário do dispositivo pode realizar esta ação.',
