--- tests/Feature/Auth/Password/ForgotPasswordAccessTest.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Tests\Feature\Auth\Password;
-
-use App\Enums\FlashMessage\FlashMessageType;
-use Illuminate\Testing\Fluent\AssertableJson;
-
-final class ForgotPasswordAccessTest extends ForgotPasswordTestSetUp
-{
-    public function test_an_unauthenticated_user_should_be_able_to_request_the_forgot_password(): void
-    {
-        $this->postJson($this->route(), [
-            'email' => $this->user->email,
-        ])
-            ->assertOk()
-            ->assertJson(fn (AssertableJson $json) => $json->where('message.type', FlashMessageType::SUCCESS)
-                ->where('message.text', trans('actions.auth.success.forgot_password')));
-    }
-}
