+++ tests/Feature/Controllers/Auth/Login/LoginTestSetUp.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Tests\Feature\Controllers\Auth\Login;
+
+use App\Models\User;
+use Database\Factories\UserFactory;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+use Tests\TestCase;
+
+class LoginTestSetUp extends TestCase
+{
+    use RefreshDatabase;
+
+    protected User $user;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $this->userSetUp();
+    }
+
+    private function userSetUp(): void
+    {
+        $this->user = UserFactory::new()->create();
+    }
+
+    protected function route(): string
+    {
+        return route('api.auth.login');
+    }
+}
