+++ tests/Feature/Controllers/Auth/Logout/LogoutAccessTest.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Tests\Feature\Controllers\Auth\Logout;
+
+use App\Enums\FlashMessage\FlashMessageType;
+use Illuminate\Testing\Fluent\AssertableJson;
+use Laravel\Sanctum\Sanctum;
+
+class LogoutAccessTest extends LogoutTestSetUp
+{
+    public function test_the_user_must_be_logged_out_and_the_access_token_destroyed(): void
+    {
+        Sanctum::actingAs($this->user);
+        $this->user->withAccessToken($this->token->accessToken);
+
+        $this->deleteJson(route('api.auth.logout'))
+            ->assertOk()
+            ->assertJson(fn (AssertableJson $json) => $json->where('message.type', FlashMessageType::SUCCESS)
+                ->where('message.text', trans('actions.auth.success.logout'))
+            );
+
+        $this->assertCount(0, $this->user->fresh()->tokens);
+    }
+}
